requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-telegram-bot==20.7
apscheduler==3.10.4
httpx~=0.25.2
pyyaml==6.0.1
python-multipart==0.0.6
